requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-OAuthlib
Flask-JWT-Extended
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
requests-oauthlib
PyJWT